Use Long for ID's instead of Int's (#263)
Room DB uses Long for number ID's.

+++ app/src/main/java/net/hawkins/gamescore/data/source/IdableDataSource.kt
@@ -4,7 +4,7 @@ import net.hawkins.gamescore.data.model.Idable
 
 interface IdableDataSource <T : Idable> {
     fun getAll(): List<T>
-    fun getById(id: Int): T?
+    fun getById(id: Long): T?
     fun save(item: T): T
-    fun deleteById(id: Int)
+    fun deleteById(id: Long)
 }
